move out network related functions
