+++ lib/constants.js
@@ -11,10 +11,6 @@
  * limitations under the License.
  */
 
-// Entry point for client utilities
-
-import * as helpers from './helpers'
-import * as proofs from './proofs'
-import * as network from './network'
-
-export default { ...helpers, ...proofs, ...network }
+// NETWORK CONSTANTS
+export const NODE_PROXY_URI = 'https://node-proxy.chainpoint.org:443'
+export const DNS_CORE_DISCOVERY_ADDR = '_core.addr.chainpoint.org'
