+++ lib/utils/helpers.js
@@ -1,52 +1,7 @@
-import url from 'url'
+import fs from 'fs'
+import crypto from 'crypto'
 import uuidValidate from 'uuid-validate'
-import { isURL, isIP } from 'validator'
-import { isEmpty, isString, has, isObject } from 'lodash'
-
-/**
- * Check if valid Core URI
- *
- * @param {string} coreURI - The Core URI to test for validity
- * @returns {bool} true if coreURI is a valid Core URI, otherwise false
- */
-export function isValidCoreURI(coreURI) {
-  if (isEmpty(coreURI) || !isString(coreURI)) return false
-
-  try {
-    return isURL(coreURI, {
-      protocols: ['https'],
-      require_protocol: true,
-      host_whitelist: [/^[a-z]\.chainpoint\.org$/]
-    })
-  } catch (error) {
-    return false
-  }
-}
-
-/**
- * Check if valid Node URI
- *
- * @param {string} nodeURI - The value to check
- * @returns {bool} true if value is a valid Node URI, otherwise false
- */
-export function isValidNodeURI(nodeURI) {
-  if (!isString(nodeURI)) return false
-
-  try {
-    let isValidURI = isURL(nodeURI, {
-      protocols: ['http', 'https'],
-      require_protocol: true,
-      host_blacklist: ['0.0.0.0']
-    })
-
-    let parsedURI = url.parse(nodeURI).hostname
-
-    // Valid URI w/ IPv4 address?
-    return isValidURI && isIP(parsedURI, 4)
-  } catch (error) {
-    return false
-  }
-}
+import { isEmpty, has, isObject } from 'lodash'
 
 /**
  * Checks if value is a hexadecimal string
@@ -90,6 +45,31 @@ export function isValidUUID(uuid) {
  */
 export function isSecureOrigin() {
   return typeof window === 'object' && window.location.protocol === 'https:'
+}
+
+export function sha256FileByPath(path) {
+  return new Promise((resolve, reject) => {
+    let sha256 = crypto.createHash('sha256')
+    let readStream = fs.createReadStream(path)
+    readStream.on('data', data => sha256.update(data))
+    readStream.on('end', () => {
+      let hash = sha256.digest('hex')
+      resolve({
+        path,
+        hash
+      })
+    })
+    readStream.on('error', err => {
+      if (err.code === 'EACCES') {
+        resolve({
+          path: path,
+          hash: null,
+          error: 'EACCES'
+        })
+      }
+      reject(err)
+    })
+  })
 }
 
 export function promiseMap(arr, fn) {
